+++ lib/image_properties.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+@immutable
+class ImageProperties {
+  final String semanticLabel;
+  final bool excludeFromSemantics;
+  final double width;
+  final double height;
+  final Color color;
+  final BlendMode colorBlendMode;
+  final BoxFit fit;
+  final AlignmentGeometry alignment;
+  final ImageRepeat repeat;
+  final Rect centerSlice;
+  final bool matchTextDirection;
+  final FilterQuality filterQuality;
+  final double scale;
+
+  const ImageProperties({
+    this.scale = 1,
+    this.semanticLabel,
+    this.excludeFromSemantics = false,
+    this.width,
+    this.height,
+    this.color,
+    this.colorBlendMode,
+    this.fit,
+    this.alignment = Alignment.center,
+    this.repeat = ImageRepeat.noRepeat,
+    this.centerSlice,
+    this.matchTextDirection = false,
+    this.filterQuality = FilterQuality.low,
+  });
+}