+++ test/html_parser_test.dart
@@ -13,6 +13,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "",
+            useRichText: false,
           ),
         ),
       ),
@@ -41,6 +42,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "<a href='https://github.com'>Test link</a>",
+            useRichText: false,
           ),
         ),
       ),
@@ -75,6 +77,7 @@ void main() {
             onLinkTap: (url) {
               urlTapped = url;
             },
+            useRichText: false,
           ),
         ),
       ),
@@ -113,6 +116,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "<abbr>Abbreviation</abbr>",
+            useRichText: false,
           ),
         ),
       ),
@@ -140,6 +144,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "<acronym>Acronym</acronym>",
+            useRichText: false,
           ),
         ),
       ),
@@ -167,6 +172,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "<address>Address</address>",
+            useRichText: false
           ),
         ),
       ),
@@ -175,18 +181,18 @@ void main() {
     expect(find.text("Address"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: "<address>Address</address>",
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: "<address>Address</address>",
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `article` tag renders", (tester) async {
@@ -195,6 +201,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: "<article>Article</article>",
+            useRichText: false,
           ),
         ),
       ),
@@ -224,6 +231,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -232,18 +240,18 @@ void main() {
     expect(find.text("Aside"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: html,
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: html,
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `b` tag renders", (tester) async {
@@ -254,6 +262,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -283,6 +292,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -291,18 +301,18 @@ void main() {
     expect(find.text("Bdi"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: html,
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: html,
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `bdo` tag renders", (tester) async {
@@ -313,6 +323,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -343,6 +354,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -351,18 +363,18 @@ void main() {
     expect(find.text("Big"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: html,
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: html,
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `blockquote` tag renders", (tester) async {
@@ -373,6 +385,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -402,6 +415,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -431,6 +445,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -461,6 +476,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -490,6 +506,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -498,18 +515,18 @@ void main() {
     expect(find.text("Cite"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: html,
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: html,
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `code` tag renders", (tester) async {
@@ -520,6 +537,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -549,6 +567,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -557,18 +576,18 @@ void main() {
     expect(find.text("Data"), findsOneWidget);
 
     //Not supported in `RichText` parser.
-//    await tester.pumpWidget(
-//      MaterialApp(
-//        home: Scaffold(
-//          body: Html(
-//            data: html,
-//            useRichText: true,
-//          ),
-//        ),
-//      ),
-//    );
-//
-//    expect(find.byType(RichText), findsOneWidget);
+    await tester.pumpWidget(
+      MaterialApp(
+        home: Scaffold(
+          body: Html(
+            data: html,
+            useRichText: true,
+          ),
+        ),
+      ),
+    );
+
+    expect(find.byType(RichText), findsOneWidget);
   });
 
   testWidgets("Check that `dd` tag renders", (tester) async {
@@ -579,6 +598,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -600,7 +620,7 @@ void main() {
     expect(find.byType(RichText), findsOneWidget);
   });
 
-  testWidgets("Check that `del` tag renders", (tester) async {
+  testWidgets("Check that `dfn` tag renders", (tester) async {
     String html = "<dfn>Dfn</dfn>";
 
     await tester.pumpWidget(
@@ -608,6 +628,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -637,6 +658,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -666,6 +688,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -695,6 +718,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -724,6 +748,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -753,6 +778,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -782,6 +808,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -811,6 +838,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -840,6 +868,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -869,6 +898,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -898,6 +928,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -927,6 +958,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -956,6 +988,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -985,6 +1018,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -1014,6 +1048,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -1043,6 +1078,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),
@@ -1072,6 +1108,7 @@ void main() {
         home: Scaffold(
           body: Html(
             data: html,
+            useRichText: false,
           ),
         ),
       ),